requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
neo4j==5.14.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
flask==2.3.3
